Tratando exception do delete

+++ src/main/java/com/estudojava/projeto_spring_boot/services/UserService.java
@@ -3,10 +3,12 @@ package com.estudojava.projeto_spring_boot.services;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.estudojava.projeto_spring_boot.entities.User;
 import com.estudojava.projeto_spring_boot.repositories.UserRepository;
+import com.estudojava.projeto_spring_boot.services.exceptions.DatabaseException;
 import com.estudojava.projeto_spring_boot.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -23,19 +25,28 @@ public class UserService {
 
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
-		//return obj.get();
+		// return obj.get();
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
-	public void delete (Long id) {
-		repository.deleteById(id);
+
+	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+
 	}
-	
-	public User update (Long id, User obj) {
+
+	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
 		return repository.save(entity);
@@ -46,5 +57,5 @@ public class UserService {
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
 	}
-	
+
 }
